requirements.txt
Pillow==11.0.0
PyYAML==6.0.2
accelerate==1.10.1
diffusers==0.35.1
einops==0.8.1
matplotlib==3.10.6
numpy==2.2.6
packaging==25.0
safetensors==0.6.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.57.0
gradio
huggingface_hub